05_databases/pyproject.toml
[project]
name = "05-databases"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
]
